+++ packages/template-retail-react-app/app/components/reset-password/index.jsx
@@ -16,64 +16,99 @@ import ResetPasswordFields from '@salesforce/retail-react-app/app/components/for
 const ResetPasswordForm = ({submitForm, clickSignIn = noop, form}) => {
     return (
         <Fragment>
-            <Stack justify="center" align="center" spacing={8}>
-                <BrandLogo width="60px" height="auto" />
-                <Stack spacing={2}>
-                    <Text align="center" fontSize="xl" fontWeight="semibold">
-                        <FormattedMessage
-                            defaultMessage="Reset Password"
-                            id="reset_password_form.title.reset_password"
-                        />
-                    </Text>
-                    <Text fontSize="sm" align="center" color="gray.700">
-                        <FormattedMessage
-                            defaultMessage="Enter your email to receive instructions on how to reset your password"
-                            id="reset_password_form.message.enter_your_email"
-                        />
-                    </Text>
-                </Stack>
-            </Stack>
-            <form onSubmit={form.handleSubmit(submitForm)} data-testid="sf-auth-modal-form">
-                <Stack paddingTop={8} spacing={8} paddingLeft={4} paddingRight={4}>
-                    {form.formState.errors?.global && (
-                        <Alert status="error">
-                            <AlertIcon color="red.500" boxSize={4} />
-                            <Text fontSize="sm" ml={3}>
-                                {form.formState.errors.global.message}
-                            </Text>
-                        </Alert>
-                    )}
-                    <ResetPasswordFields form={form} />
-                    <Stack spacing={6}>
-                        <Button
-                            type="submit"
-                            onClick={() => form.clearErrors('global')}
-                            isLoading={form.formState.isSubmitting}
-                        >
-                            <FormattedMessage
-                                defaultMessage="Reset Password"
-                                id="reset_password_form.button.reset_password"
-                            />
-                        </Button>
-
-                        <Stack direction="row" spacing={1} justify="center">
-                            <Text fontSize="sm">
+            {!form.formState.isSubmitSuccessful && (
+                <>
+                    <Stack justify="center" align="center" spacing={8}>
+                        <BrandLogo width="60px" height="auto" />
+                        <Stack spacing={2}>
+                            <Text align="center" fontSize="xl" fontWeight="semibold">
                                 <FormattedMessage
-                                    defaultMessage="Or return to"
-                                    id="reset_password_form.message.return_to_sign_in"
-                                    description="Precedes link to return to sign in"
+                                    defaultMessage="Reset Password"
+                                    id="reset_password_form.title.reset_password"
                                 />
                             </Text>
-                            <Button variant="link" size="sm" onClick={clickSignIn}>
+                            <Text fontSize="sm" align="center" color="gray.700">
                                 <FormattedMessage
-                                    defaultMessage="Sign in"
-                                    id="reset_password_form.action.sign_in"
+                                    defaultMessage="Enter your email to receive instructions on how to reset your password"
+                                    id="reset_password_form.message.enter_your_email"
                                 />
-                            </Button>
+                            </Text>
                         </Stack>
                     </Stack>
+                    <form onSubmit={form.handleSubmit(submitForm)} data-testid="sf-auth-modal-form">
+                        <Stack paddingTop={8} spacing={8} paddingLeft={4} paddingRight={4}>
+                            {form.formState.errors?.global && (
+                                <Alert status="error">
+                                    <AlertIcon color="red.500" boxSize={4} />
+                                    <Text fontSize="sm" ml={3}>
+                                        {form.formState.errors.global.message}
+                                    </Text>
+                                </Alert>
+                            )}
+                            <ResetPasswordFields form={form} />
+                            <Stack spacing={6}>
+                                <Button
+                                    type="submit"
+                                    onClick={() => form.clearErrors('global')}
+                                    isLoading={form.formState.isSubmitting}
+                                >
+                                    <FormattedMessage
+                                        defaultMessage="Reset Password"
+                                        id="reset_password_form.button.reset_password"
+                                    />
+                                </Button>
+
+                                <Stack direction="row" spacing={1} justify="center">
+                                    <Text fontSize="sm">
+                                        <FormattedMessage
+                                            defaultMessage="Or return to"
+                                            id="reset_password_form.message.return_to_sign_in"
+                                            description="Precedes link to return to sign in"
+                                        />
+                                    </Text>
+                                    <Button variant="link" size="sm" onClick={clickSignIn}>
+                                        <FormattedMessage
+                                            defaultMessage="Sign in"
+                                            id="reset_password_form.action.sign_in"
+                                        />
+                                    </Button>
+                                </Stack>
+                            </Stack>
+                        </Stack>
+                    </form>
+                </>
+            )}
+            {form.formState.isSubmitSuccessful && (
+                <Stack justify="center" align="center" spacing={6}>
+                    <BrandLogo width="60px" height="auto" />
+                    <Text align="center" fontSize="md">
+                        <FormattedMessage
+                            defaultMessage={'Password Reset'}
+                            id="auth_modal.password_reset_success.title.password_reset"
+                        />
+                    </Text>
+                    <Stack spacing={6} pt={4}>
+                        <Text align="center" fontSize="sm">
+                            <FormattedMessage
+                                defaultMessage="You will receive an email at <b>{email}</b> with a link to reset your password shortly."
+                                id="auth_modal.password_reset_success.info.will_email_shortly"
+                                values={{
+                                    email: form.getValues('email'),
+
+                                    b: (chunks) => <b>{chunks}</b>
+                                }}
+                            />
+                        </Text>
+
+                        <Button onClick={clickSignIn}>
+                            <FormattedMessage
+                                defaultMessage="Back to Sign In"
+                                id="auth_modal.password_reset_success.button.back_to_sign_in"
+                            />
+                        </Button>
+                    </Stack>
                 </Stack>
-            </form>
+            )}
         </Fragment>
     )
 }
