+++ packages/template-retail-react-app/translations/en-US.json
@@ -627,6 +627,9 @@
   "global.account.link.wishlist": {
     "defaultMessage": "Wishlist"
   },
+  "global.error.invalid_token": {
+    "defaultMessage": "Invalid token, please try again."
+  },
   "global.error.something_went_wrong": {
     "defaultMessage": "Something went wrong. Try again!"
   },
@@ -977,9 +980,6 @@
   },
   "login_page.error.incorrect_username_or_password": {
     "defaultMessage": "Incorrect username or password, please try again."
-  },
-  "login_page.error.invalid_token": {
-    "defaultMessage": "Invalid token, please try again."
   },
   "offline_banner.description.browsing_offline_mode": {
     "defaultMessage": "You're currently browsing in offline mode"
